requirements.test.txt
factory-boy==3.2.1
freezegun==1.1.0
pytest==8.3.3
pytest-django==4.6.0
